correçõe dos bugs da aplicação, implementando lógica de exeptions, reformulando todo html e css

+++ src/main/java/com/RODRIGO/RPX/services/MarcaService.java
@@ -2,6 +2,7 @@ package com.RODRIGO.RPX.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.RODRIGO.RPX.entity.Marca;
 import com.RODRIGO.RPX.repository.MarcaRepository;
@@ -14,7 +15,15 @@ public class MarcaService {
     {
         return marcaRepository.findById(id).orElseThrow(() -> new RuntimeException("Produto não encontrado"));
     }
+    @Transactional
     public void deletar(Long id){
-        marcaRepository.deleteById(id);
+        Marca marca = marcaRepository.findById(id)
+        .orElseThrow(() -> new RuntimeException("Marca não encontrada"));
+
+    if (!marca.getProdutos().isEmpty()) {
+        throw new RuntimeException("Não é possível deletar: existem produtos associados a esta marca.");
+    }
+
+    marcaRepository.deleteById(id);
     }
 }
